Move request buttons tests to separate file
